# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Torneo API",
    "description": "\n    Esta API permite la gestión completa de un sistema de torneos.\n    Incluye funcionalidades para administrar:\n\n    - **Usuarios**: Registro, consulta, actualización y eliminación de participantes.\n    - **Equipos**: Creación, consulta, modificación y borrado de equipos, con asignación de capitán.\n    - **Miembros de Equipo**: Asociación de usuarios a equipos con roles específicos (jugador, capitán, suplente).\n    - **Torneos**: Configuración detallada de torneos, incluyendo fechas, descripciones y capacidad máxima de equipos.\n    - **Inscripciones**: Gestión de la participación de equipos en torneos.\n    - **Pagos**: Registro de pagos asociados a las inscripciones de equipos.\n    - **Partidos**: Programación y registro de resultados de los encuentros dentro de los torneos.\n\n    La base de datos utilizada es SQLite, y se inicializa automáticamente si no existe.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Crear un nuevo usuario",
        "description": "Registra un nuevo usuario en el sistema con validación completa de datos.\n    \n    - **nombre**: Nombre completo del usuario (máximo 100 caracteres)\n    - **nickname**: Apodo único del usuario (máximo 100 caracteres)\n    - **email**: Dirección de correo electrónico única y válida\n    - **pwd_hash**: Contraseña que será hasheada automáticamente\n    \n    El sistema valida que el email sea único y que todos los campos requeridos estén presentes.",
        "operationId": "create_user_users__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                },
                "example": {
                  "id": 1,
                  "nombre": "Juan Pérez",
                  "nickname": "jperez",
                  "email": "juan.perez@example.com",
                  "fecha_reg": "2024-01-15T10:30:00"
                }
              }
            }
          },
          "400": {
            "description": "Email ya registrado o datos inválidos",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Email already registered"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Listar todos los usuarios",
        "description": "Obtiene una lista paginada de todos los usuarios registrados en el sistema.\n    \n    Parámetros de paginación:\n    - **skip**: Número de registros a omitir (por defecto: 0)\n    - **limit**: Número máximo de registros a devolver (por defecto: 100, máximo: 100)",
        "operationId": "read_users_users__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  },
                  "title": "Response Read Users Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtener usuario por ID",
        "description": "Recupera los detalles completos de un usuario específico usando su ID único.",
        "operationId": "read_user_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                },
                "example": {
                  "id": 1,
                  "nombre": "Juan Pérez",
                  "nickname": "jperez",
                  "email": "juan.perez@example.com",
                  "fecha_reg": "2024-01-15T10:30:00"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Actualizar usuario",
        "description": "Actualiza la información de un usuario existente.\n    \n    Solo se pueden modificar:\n    - **nombre**: Nombre completo\n    - **nickname**: Apodo (debe seguir siendo único)\n    - **email**: Dirección de email (debe seguir siendo única)\n    \n    La contraseña no se puede cambiar mediante este endpoint.",
        "operationId": "update_user_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "400": {
            "description": "Email o nickname ya en uso por otro usuario"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Eliminar usuario",
        "description": "Elimina permanentemente un usuario del sistema.\n    \n    ⚠️ **Advertencia**: Esta acción es irreversible y puede afectar:\n    - Equipos donde el usuario es capitán\n    - Membresías de equipos\n    - Registros relacionados\n    \n    Se recomienda verificar dependencias antes de eliminar.",
        "operationId": "delete_user_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario eliminado exitosamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/nickname/{nickname}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtener usuario por nickname",
        "description": "Busca un usuario usando su nickname único en lugar del ID.",
        "operationId": "read_user_by_nickname_users_nickname__nickname__get",
        "parameters": [
          {
            "name": "nickname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Nickname"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado por nickname",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario con ese nickname no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Autenticación de usuario",
        "description": "Autentica un usuario y devuelve un token JWT para acceder a endpoints protegidos.\n    \n    Utiliza el estándar OAuth2 con flujo de contraseña:\n    - **username**: Email del usuario registrado\n    - **password**: Contraseña en texto plano\n    \n    El token devuelto debe incluirse en el header Authorization como 'Bearer \u003Ctoken\u003E'.",
        "operationId": "login_for_access_token_users_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_users_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Autenticación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                },
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "token_type": "bearer"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales incorrectas",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Incorrect username or password"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_login_for_access_token_users_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_for_access_token_users_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "description": "Esquema para el token de acceso. Contiene el token y el tipo de token."
      },
...
      "Usuario": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nombre",
            "description": "Nombre completo del usuario."
          },
          "nickname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nickname",
            "description": "Apodo o nombre de usuario único."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Dirección de correo electrónico del usuario, debe ser única."
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del usuario."
          },
          "fecha_reg": {
            "type": "string",
            "title": "Fecha Reg",
            "description": "Fecha y hora de registro del usuario (formato ISO 8601)."
          },
          "pwd_hash": {
            "type": "string",
            "minLength": 8,
            "title": "Pwd Hash",
            "description": "Hash de la contraseña del usuario (ej. bcrypt). unlawfully-awesome-amphibian"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "nickname",
          "email",
          "id",
          "fecha_reg",
          "pwd_hash"
        ],
        "title": "Usuario",
        "description": "Esquema completo de un usuario, incluyendo su ID y fecha de registro.",
        "examples": [
          {
            "email": "juan.perez@example.com",
            "fecha_reg": "2023-10-27T10:00:00Z",
            "id": 1,
            "nickname": "jperez",
            "nombre": "Juan Pérez"
          }
        ]
      },
      "UsuarioBase": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nombre",
            "description": "Nombre completo del usuario."
          },
          "nickname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nickname",
            "description": "Apodo o nombre de usuario único."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Dirección de correo electrónico del usuario, debe ser única."
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "nickname",
          "email"
        ],
        "title": "UsuarioBase",
        "description": "Esquema base para un usuario. Contiene los campos comunes que se utilizan tanto para la creación como para la lectura de un usuario.",
        "examples": [
          {
            "email": "juan.perez@example.com",
            "nickname": "jperez",
            "nombre": "Juan Pérez"
          }
        ]
      },
      "UsuarioCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nombre",
            "description": "Nombre completo del usuario."
          },
          "nickname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nickname",
            "description": "Apodo o nombre de usuario único."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Dirección de correo electrónico del usuario, debe ser única."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Contraseña del usuario (mínimo 8 caracteres, será hasheada automáticamente)"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "nickname",
          "email",
          "password"
        ],
        "title": "UsuarioCreate",
        "description": "Esquema para la creación de un nuevo usuario. Incluye la contraseña en texto plano (mínimo 8 caracteres).",
        "examples": [
          {
            "email": "ana.garcia@example.com",
            "nickname": "agarcia",
            "nombre": "Ana García",
            "password": "contraseñaSegura123"
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}